track qpu executions and shots on default qubit
