+++ tests/devices/experimental/test_default_qubit_2.py
@@ -20,7 +20,7 @@ import numpy as np
 
 import pennylane as qml
 from pennylane.measurements import SampleMP, StateMP, ProbabilityMP
-from pennylane.resource import Resources
+
 from pennylane.devices import DefaultQubit, ExecutionConfig
 from pennylane.devices.qubit.preprocess import validate_and_expand_adjoint
 
@@ -96,159 +96,6 @@ class TestSnapshotMulti:
             match="Debugging with ``Snapshots`` is not available with multiprocessing.",
         ):
             qml.snapshots(circuit)()
-
-
-class TestTracking:
-    """Testing the tracking capabilities of DefaultQubit."""
-
-    def test_tracker_set_upon_initialization(self):
-        """Test that a new tracker is intialized with each device."""
-        assert DefaultQubit().tracker is not DefaultQubit().tracker
-
-    def test_tracker_not_updated_if_not_active(self):
-        """Test that the tracker is not updated if not active."""
-        dev = DefaultQubit()
-        assert len(dev.tracker.totals) == 0
-
-        dev.execute(qml.tape.QuantumScript())
-        assert len(dev.tracker.totals) == 0
-        assert len(dev.tracker.history) == 0
-
-    def test_tracking_batch(self):
-        """Test that the new default qubit integrates with the tracker."""
-
-        qs = qml.tape.QuantumScript([], [qml.expval(qml.PauliZ(0))])
-
-        dev = DefaultQubit()
-        config = ExecutionConfig(gradient_method="adjoint")
-        with qml.Tracker(dev) as tracker:
-            dev.execute(qs)
-            dev.compute_derivatives(qs, config)
-            dev.execute([qs, qs])  # and a second time
-
-        assert tracker.history == {
-            "batches": [1, 1],
-            "executions": [1, 2],
-            "resources": [Resources(num_wires=1), Resources(num_wires=1), Resources(num_wires=1)],
-            "derivative_batches": [1],
-            "derivatives": [1],
-        }
-        assert tracker.totals == {
-            "batches": 2,
-            "executions": 3,
-            "derivative_batches": 1,
-            "derivatives": 1,
-        }
-        assert tracker.latest == {"batches": 1, "executions": 2}
-
-    def test_tracking_execute_and_derivatives(self):
-        """Test that the execute_and_compute_* calls are being tracked for the
-        new default qubit device"""
-
-        qs = qml.tape.QuantumScript([], [qml.expval(qml.PauliZ(0))])
-        dev = DefaultQubit()
-        config = ExecutionConfig(gradient_method="adjoint")
-
-        with qml.Tracker(dev) as tracker:
-            dev.compute_derivatives(qs, config)
-            dev.execute_and_compute_derivatives([qs] * 2, config)
-            dev.compute_jvp([qs] * 3, [(0,)] * 3, config)
-            dev.execute_and_compute_jvp([qs] * 4, [(0,)] * 4, config)
-            dev.compute_vjp([qs] * 5, [(0,)] * 5, config)
-            dev.execute_and_compute_vjp([qs] * 6, [(0,)] * 6, config)
-
-        assert tracker.history == {
-            "executions": [2, 4, 6],
-            "derivatives": [1, 2],
-            "derivative_batches": [1],
-            "execute_and_derivative_batches": [1],
-            "jvps": [3, 4],
-            "jvp_batches": [1],
-            "execute_and_jvp_batches": [1],
-            "vjps": [5, 6],
-            "vjp_batches": [1],
-            "execute_and_vjp_batches": [1],
-            "resources": [Resources(num_wires=1)] * 12,
-        }
-
-    def test_tracking_resources(self):
-        """Test that resources are tracked for the new default qubit device."""
-        qs = qml.tape.QuantumScript(
-            [
-                qml.Hadamard(0),
-                qml.Hadamard(1),
-                qml.CNOT(wires=[0, 2]),
-                qml.RZ(1.23, 1),
-                qml.CNOT(wires=[1, 2]),
-                qml.Hadamard(0),
-            ],
-            [qml.expval(qml.PauliZ(1)), qml.expval(qml.PauliY(2))],
-        )
-
-        expected_resources = Resources(
-            num_wires=3,
-            num_gates=6,
-            gate_types={"Hadamard": 3, "CNOT": 2, "RZ": 1},
-            gate_sizes={1: 4, 2: 2},
-            depth=3,
-        )
-
-        dev = DefaultQubit()
-        with qml.Tracker(dev) as tracker:
-            dev.execute(qs)
-
-        assert len(tracker.history["resources"]) == 1
-        assert tracker.history["resources"][0] == expected_resources
-
-    def test_tracking_batched_execution(self):
-        """Test the number of times the device is executed over a QNode's
-        lifetime is tracked by the device's tracker."""
-
-        dev_1 = qml.device("default.qubit", wires=2)
-
-        def circuit_1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node_1 = qml.QNode(circuit_1, dev_1)
-        num_evals_1 = 10
-
-        with qml.Tracker(dev_1, persistent=True) as tracker1:
-            for _ in range(num_evals_1):
-                node_1(0.432, np.array([0.12, 0.5, 3.2]))
-        assert tracker1.totals["executions"] == num_evals_1
-
-        # test a second instance of a default qubit device
-        dev_2 = qml.device("default.qubit", wires=2)
-
-        def circuit_2(x):
-            qml.RX(x, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node_2 = qml.QNode(circuit_2, dev_2)
-        num_evals_2 = 5
-
-        with qml.Tracker(dev_2) as tracker2:
-            for _ in range(num_evals_2):
-                node_2(np.array([0.432, 0.61, 8.2]))
-        assert tracker2.totals["executions"] == num_evals_2
-
-        # test a new circuit on an existing instance of a qubit device
-        def circuit_3(y):
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node_3 = qml.QNode(circuit_3, dev_1)
-        num_evals_3 = 7
-
-        with tracker1:
-            for _ in range(num_evals_3):
-                node_3(np.array([0.12, 1.214]))
-        assert tracker1.totals["executions"] == num_evals_1 + num_evals_3
 
 
 # pylint: disable=too-few-public-methods
